Updated the Asus info.
I emailed Asus and got a response saying that they are aware of the
problem and are working to fix it. They also gave me a list of routers
that are not affected by this issue in Router or AP (Access Point) mode.

Since they are both working on it, and there is the potential for no
issues for some people, I checked off two boxes in the first chart.